test(physics): relocate VENUS USR fixture tests + correct PLEIADES misnomer (#497)
Move 13 fixture-gated `#[ignore]`'d tests out of `crates/nereids-physics/src/`
into proper integration tests under `tests/`, and correct the misnomer that
labeled the fixture as "PLEIADES" (it is a SAMMY-format tabulated resolution
kernel for the VENUS instrument).

Relocation:
  - 8 regression tests + 1 bit-exact test
    → tests/venus_usr_resolution.rs
  - 3 microbenchmarks
    → tests/venus_usr_resolution_microbench.rs
  - 2 surrogate-cubature regressions
    → tests/venus_usr_surrogate.rs

Tests now use the early-return idiom on `common::venus_usr_resolution_path()`
rather than `#[ignore]` — when the fixture is present they execute, when
absent they pass as no-ops, removing the `0 passed; N ignored` noise from
`cargo test`.

Public-API surface minimized: integration tests rely on existing `pub` API
(`apply_r`, `apply_resolution_with_matrix`, `compile_to_matrix`, `plan`,
`from_text`, the `SparseEmpiricalCubaturePlan` / `ScalarChebyshevPlan` types)
plus three thin shims added to the existing `resolution::test_support` module
gated behind the existing `test-support` feature:
  - `broaden_presorted` — exposes the crate-internal optimized two-pointer walk
  - `interpolated_kernel` — needed by the bit-exact oracle in the integration
    test to share interior math with the SUT
  - `TOF_FACTOR` const — same rationale: shared constant for bit-exactness

`build_fixture_plan_and_matrix`, `broaden_presorted_reference`,
`assert_bit_exact`, `max_hybrid_err`, `exact_forward`, `interp_spectrum`
are duplicated into the integration tests rather than promoted — they are
short helpers and duplication keeps the public surface tight.

`crates/nereids-physics/Cargo.toml` gains a `dev-dependencies` self-link with
the `test-support` feature enabled so integration tests can see the shims
without leaking them into release builds.

The fixture file `_fts_bl10_0p5meV_1keV_25pts.txt` is gitignored per ORNL
release policy and remains so.  PLEIADES references in `crates/nereids-io/`
(algorithm citations for ORNL Method 2 normalization) are legitimate and
were not touched.

Validation:
  - cargo fmt clean
  - cargo clippy --workspace --exclude nereids-python --all-targets -- -D warnings clean
  - cargo test --workspace --exclude nereids-python: all pass
  - fixture present: 13 relocated tests execute and pass
  - fixture absent: 13 tests pass as no-ops in <1ms total

Closes #497.
